misoul-api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.3.1
langchain==0.0.267
langchain_community==0.0.16

transformers==4.36.2
sentence-transformers==2.2.2
qdrant-client==1.6.4
python-dotenv==1.0.0
gunicorn==20.1.0